services/hass_helper/requirements.txt
fastapi>=0.103,<1
uvicorn[standard]>=0.23,<1
httpx>=0.24,<1
python-dotenv>=0.21,<1
pydantic>=1.10,<3
